+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Travis-CI integration
 - Codecov integration
 - Bettercodehub integration
+- translator submodule
+- Kapellmeister class
+
+### Changed
+- Kapellmeister is now used in main.py instead of Runner
 
 ## 0.1.0 - 2018-08-19
 ### Added